requirements.txt
mss
opencv-python
numpy
requests
pyyaml
pillow
